requirements.txt
streamlit
anthropic
jira
python-dotenv
